ElasticStack_graph_apache/requirements.txt
python-dateutil==2.6.0
requests==2.12.1
six==1.10.0
